Allow specification of IP address for service
The current address_for method requires that a service pass in the
interface that it wishes to bind to.  It then looks up the first
IP address on the interface.  This does not work when there are
multiple IP addresses assigned to the interface.  This change adds
the address_for_service method that allows a service to specify the
IP address to bind to.  If no IP address is specified then this
method will fall back to calling the address_for method.